Disable description parsing for now
Atlassian have changed the format in which they return issue
descriptions now. Instead of the old ADF formatted text string, we
instead get a nested JSON data structure with markup elements. While
that's better in principle, since ADF is weird and nonstandard, it'll
take a bit more than just piping it through pandoc now in order to
parse it.